fix(db-dynamodb): remove unnecessary export
