+++ packages/db-dynamodb/src/utils/batch/index.ts
@@ -1,8 +1,3 @@
 export * from "./batchRead";
 export * from "./batchWrite";
-export * from "./EntityReadBatch";
-export * from "./EntityReadBatchBuilder";
-export * from "./EntityWriteBatch";
-export * from "./EntityWriteBatchBuilder";
-export * from "./TableWriteBatch";
 export * from "./types";
