+++ packages/db-dynamodb/src/utils/entity/index.ts
@@ -1,2 +1,6 @@
 export * from "./Entity";
+export * from "./EntityReadBatch";
+export * from "./EntityReadBatchBuilder";
+export * from "./EntityWriteBatch";
+export * from "./EntityWriteBatchBuilder";
 export * from "./types";
