+++ packages/db-dynamodb/src/utils/index.ts
@@ -11,7 +11,6 @@ export * from "./query";
 export * from "./count";
 export * from "./scan";
 export * from "./sort";
-export * from "./table";
 export * from "./update";
 export * from "./batch";
 export * from "./entity";
