+++ packages/db-dynamodb/src/utils/table/index.ts
@@ -1,2 +1,4 @@
 export * from "./Table";
+export * from "./TableReadBatch";
+export * from "./TableWriteBatch";
 export * from "./types";
